fix(test+ci): wire mocked playwright coverage, expand audit, drop skipped spec
Follow-up to the previous review pass on #1228; addresses the remaining
gaps Claude (Opus 4.7) flagged in code review.

Audit:
- src/server/api/mod.rs: add api/log_level.rs::patch_log_level to the
  read-only audit handler list. It already has a state.read_only guard,
  but was not previously enforced by the audit; a future removal would
  go unnoticed.

Mocked Playwright coverage:
- web/tests/helpers/coverageCapture.ts: extract the shared
  window.__coverage__ capture helper used by liveTest.ts so it can be
  reused by the mocked suite.
- web/tests/helpers/mockedTest.ts: thin Playwright test wrapper that
  overrides the page fixture with the same capture hook. Specs under
  web/tests/*.spec.ts now go through this wrapper.
- Update all 25 mocked specs to import { test, expect } from
  ./helpers/mockedTest. Auxiliary type imports (Page, devices) still
  come from @playwright/test directly. Without this, the mocked
  Playwright CI job was uploading a near-empty LCOV to Codecov despite
  building with AOE_COVERAGE=1.
- web/tests/helpers/liveTest.ts: thin out to import captureCoverage
  from the shared helper. Also document the auth-mode constraint on
  the serveCockpit fixture (no-auth only today; passphrase+cockpit
  needs the harness to thread sessionCookie through to PATCH
  /api/cockpit/master).

Coverage matrix:
- web/tests/live/auth-login-passphrase.spec.ts: delete. Both tests
  were test.skip()'d pending #1230 (loopback-bypass blocks LoginPage
  from rendering under the harness's 127.0.0.1 bind). An all-skipped
  file is easy to overlook when later harness changes silently break
  the imports; reintroduce alongside the bypass fix.
- web/tests/coverage-matrix.json: flip auth.passphrase-login to
  kind: "deferred" with issue: #1230 so the surface stays tracked
  without referencing a missing spec file. validate-coverage-matrix.mjs
  passes.

Em-dash:
- web/tests/live/auth-login-passphrase.spec.ts (now deleted) had one
  em-dash in a code comment. Removed alongside the file.

Validation:
- cargo fmt --check: clean.
- cargo clippy --features serve --all-targets -- -D warnings: clean.
- cargo test --features serve --lib server::api::tests: 7/7 pass
  (including every_mutating_handler_has_read_only_guard with the new
  patch_log_level entry).
- npm run test:unit: 313 vitest tests pass.
- npx playwright test --config=playwright.config.ts --list: 148 tests
  across 25 specs discovered through the new mockedTest wrapper.
- npx playwright test --config=playwright.live.config.ts --list: 11
  tests across 8 specs (auth-login-passphrase removed).
- node web/tests/validate-coverage-matrix.mjs: passes.

Part of #986.
