+++ web/tests/helpers/liveTest.ts
@@ -21,51 +21,22 @@
 //   });
 
 import { test as base, expect, type Page } from "@playwright/test";
-import { mkdir, writeFile } from "node:fs/promises";
-import { resolve, dirname } from "node:path";
-import { fileURLToPath } from "node:url";
-import { randomUUID } from "node:crypto";
 import { spawnAoeServe, type ServeHandle } from "./aoeServe";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-const playwrightCoverageDir = resolve(
-  __dirname,
-  "..",
-  "..",
-  "coverage",
-  "playwright",
-);
+import { captureCoverage } from "./coverageCapture";
 
 type LiveFixtures = {
   serve: ServeHandle;
   servePassphrase: ServeHandle;
   serveReadOnly: ServeHandle;
+  /**
+   * Cockpit fixture. Only supported with `authMode: "none"` today; the
+   * harness calls `PATCH /api/cockpit/master` without a session cookie.
+   * Token-mode cockpit coverage is queued in #1226. If you need
+   * passphrase + cockpit, call `spawnAoeServe` directly and pass the
+   * `sessionCookie` through to the master-enable request.
+   */
   serveCockpit: ServeHandle;
 };
-
-async function captureCoverage(page: Page, testTitle: string): Promise<void> {
-  if (process.env.AOE_COVERAGE !== "1") return;
-  try {
-    const coverage = await page.evaluate(
-      () => (window as unknown as { __coverage__?: unknown }).__coverage__,
-    );
-    if (!coverage) return;
-    await mkdir(playwrightCoverageDir, { recursive: true });
-    const safe = testTitle
-      .replace(/[^a-zA-Z0-9_-]/g, "_")
-      .slice(0, 160);
-    const filename = `${safe}-${randomUUID()}.json`;
-    await writeFile(
-      resolve(playwrightCoverageDir, filename),
-      JSON.stringify(coverage),
-    );
-  } catch {
-    // Test may have closed the page or navigated cross-origin. Coverage
-    // gaps in those edge cases are acceptable; we'd rather not fail the
-    // test because of a coverage-collection hiccup.
-  }
-}
 
 /**
  * Seed the session cookie + device binding secret onto a page so a
