+++ web/tests/wheel-scroll.spec.ts
@@ -1,4 +1,5 @@
-import { test, expect, type Page } from "@playwright/test";
+import { test, expect } from "./helpers/mockedTest";
+import type { Page } from "@playwright/test";
 import { clickSidebarSession } from "./helpers/sidebar";
 import {
   mockTerminalApis,
